+++ git-packetline/src/read/async_io.rs
@@ -7,7 +7,7 @@ use futures_lite::AsyncReadExt;
 use crate::{
     decode,
     read::{ExhaustiveOutcome, WithSidebands},
-    PacketLine, StreamingPeekableIter, MAX_LINE_LEN, U16_HEX_BYTES,
+    PacketLineRef, StreamingPeekableIter, MAX_LINE_LEN, U16_HEX_BYTES,
 };
 
 /// Non-IO methods
@@ -19,7 +19,7 @@ where
     async fn read_line_inner<'a>(
         reader: &mut T,
         buf: &'a mut Vec<u8>,
-    ) -> io::Result<Result<PacketLine<'a>, decode::Error>> {
+    ) -> io::Result<Result<PacketLineRef<'a>, decode::Error>> {
         let (hex_bytes, data_bytes) = buf.split_at_mut(4);
         reader.read_exact(hex_bytes).await?;
         let num_data_bytes = match decode::hex_prefix(hex_bytes) {
@@ -41,7 +41,7 @@ where
     async fn read_line_inner_exhaustive<'a>(
         reader: &mut T,
         buf: &'a mut Vec<u8>,
-        delimiters: &[PacketLine<'static>],
+        delimiters: &[PacketLineRef<'static>],
         fail_on_err_lines: bool,
         buf_resize: bool,
     ) -> ExhaustiveOutcome<'a> {
@@ -89,7 +89,7 @@ where
     ///  * natural EOF
     ///  * ERR packet line encountered if [`fail_on_err_lines()`][StreamingPeekableIter::fail_on_err_lines()] is true.
     ///  * A `delimiter` packet line encountered
-    pub async fn read_line(&mut self) -> Option<io::Result<Result<PacketLine<'_>, decode::Error>>> {
+    pub async fn read_line(&mut self) -> Option<io::Result<Result<PacketLineRef<'_>, decode::Error>>> {
         if self.is_done {
             return None;
         }
@@ -118,7 +118,7 @@ where
     /// Peek the next packet line without consuming it.
     ///
     /// Multiple calls to peek will return the same packet line, if there is one.
-    pub async fn peek_line(&mut self) -> Option<io::Result<Result<PacketLine<'_>, decode::Error>>> {
+    pub async fn peek_line(&mut self) -> Option<io::Result<Result<PacketLineRef<'_>, decode::Error>>> {
         if self.is_done {
             return None;
         }