+++ git-transport/src/client/git/blocking_io.rs
@@ -1,7 +1,7 @@
 use std::{io, io::Write};
 
 use bstr::BString;
-use git_packetline::PacketLine;
+use git_packetline::PacketLineRef;
 
 use crate::{
     client::{self, capabilities, git, Capabilities, SetServiceResponse},
@@ -113,7 +113,7 @@ where
     ) -> Self {
         git::Connection {
             writer: write,
-            line_provider: git_packetline::StreamingPeekableIter::new(read, &[PacketLine::Flush]),
+            line_provider: git_packetline::StreamingPeekableIter::new(read, &[PacketLineRef::Flush]),
             path: repository_path.into(),
             virtual_host: virtual_host.map(|(h, p)| (h.into(), p)),
             desired_version,