+++ git-packetline/CHANGELOG.md
@@ -1,3 +1,8 @@
 ### 0.10.0 (2021-08-??)
 
-- TBD
+#### Breaking
+
+* **renames / moves**
+    - `immutable::PacketLine` -> `PacketLineRef`
+    - `immutable::DecodeBandError` -> `decode::band::Error`
+